+++ packages/rotate/src/rotatePlugin.tsx
@@ -7,14 +7,13 @@
  */
 
 import * as React from 'react';
-import { createStore } from '@react-pdf-viewer/core';
+import { createStore, RotateDirection } from '@react-pdf-viewer/core';
 import type { Plugin, PluginFunctions, ViewerState } from '@react-pdf-viewer/core';
 
 import { Rotate, RotateProps } from './Rotate';
 import { RotateButton } from './RotateButton';
 import { RotateMenuItem } from './RotateMenuItem';
 import { RotatePage, RotatePageProps } from './RotatePage';
-import { RotateDirection } from './structs/RotateDirection';
 import type { StoreProps } from './types/StoreProps';
 
 export interface RotateDecoratorProps {